+++ database/sql/src/main/scala/cromwell/database/sql/tables/DockerHashStoreEntry.scala
@@ -5,5 +5,6 @@ case class DockerHashStoreEntry
   workflowExecutionUuid: String,
   dockerTag: String,
   dockerHash: String,
+  dockerSize: Option[Long],
   dockerHashStoreEntryId: Option[Int] = None
 )