+++ dockerHashing/src/main/scala/cromwell/docker/registryv2/DockerAccessToken.scala
@@ -1,0 +1,3 @@
+package cromwell.docker.registryv2
+
+case class DockerAccessToken(token: String)